Completed 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -50,9 +50,18 @@ blogsRouter.post('/', async (request, response, next) => {
 });
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' });
+  }
+
   try {
-    await Blog.findByIdAndDelete(request.params.id);
-    response.status(204).end();
+    const blog = await Blog.findById(request.params.id);
+    if (blog.user.toString() === decodedToken.toString()) {
+      Blog.findByIdAndDelete(request.params.id);
+      response.status(204).end();
+    }
   } catch (error) {
     next(error);
   }
